requirements.txt
streamlit
google-generativeai
pymupdf
pytesseract
opencv-python
